abstracted console controller class, it is seperate frmo uesr input class and had too many responsabilities. also simulation only needs 'controls' not the capacity to initialize the program/state of the world
