+++ Conways/Program.cs
@@ -5,11 +5,12 @@
     static void Main(string[] args)
     {
       var userInput = new ConsoleInput();
+      var controller = new ConsoleController();
+      var renderer = new ConsoleRenderer();
       var dimensions = userInput.GetDimensions();
       var indexes = userInput.GetValidIndexes(dimensions);
       var world = new World(dimensions.Item1, dimensions.Item2, indexes);
-      var renderer = new ConsoleRenderer();
-      Simulation.Run(userInput, renderer, world);
+      Simulation.Run(controller, renderer, world);
     }
   }
 }
